filestats change

+++ code/go/0chain.net/blobbercore/reference/ref.go
@@ -512,46 +512,51 @@ func DeleteReference(ctx context.Context, refID int64, pathHash string) error {
 
 func (r *Ref) SaveFileRef(ctx context.Context) error {
 	db := datastore.GetStore().GetTransaction(ctx)
-	err := db.Delete(&Ref{}, "id=?", r.ID).Error
-	if err != nil && err != gorm.ErrRecordNotFound {
-		return err
-	}
-	r.IsTemp = true
-	r.ID = 0
-	return db.Create(r).Error
-}
-
-func (r *Ref) SaveDirRef(ctx context.Context) error {
-	db := datastore.GetStore().GetTransaction(ctx)
-	db = db.Model(r).Where("id = ?", r.ID).Updates(map[string]interface{}{
-		"allocation_id":      r.AllocationID,
-		"lookup_hash":        r.LookupHash,
-		"name":               r.Name,
-		"path":               r.Path,
-		"hash":               r.Hash,
-		"file_meta_hash":     r.FileMetaHash,
-		"num_of_blocks":      r.NumBlocks,
-		"path_hash":          r.PathHash,
-		"parent_path":        r.ParentPath,
-		"level":              r.PathLevel,
-		"allocation_root":    r.AllocationRoot,
-		"size":               r.Size,
-		"chunk_size":         r.ChunkSize,
-		"is_temp":            false,
-		"file_id":            r.FileID,
-		"thumbnail_filename": r.ThumbnailFilename,
-	})
-	if errors.Is(db.Error, gorm.ErrRecordNotFound) || db.RowsAffected == 0 {
+	toUpdateFileStat := r.IsTemp
+	prevID := r.ID
+	if r.ID > 0 {
 		err := db.Delete(&Ref{}, "id=?", r.ID).Error
 		if err != nil && err != gorm.ErrRecordNotFound {
 			return err
 		}
-		r.IsTemp = true
-		r.ID = 0
-		return db.Create(r).Error
-	} else {
-		return db.Error
 	}
+	r.IsTemp = true
+	r.ID = 0
+	err := db.Create(r).Error
+	if err != nil {
+		return err
+	}
+
+	if toUpdateFileStat {
+		FileUpdated(ctx, prevID, r.ID)
+	}
+
+	return nil
+}
+
+func (r *Ref) SaveDirRef(ctx context.Context) error {
+	db := datastore.GetStore().GetTransaction(ctx)
+	toUpdateFileStat := r.IsTemp
+	prevID := r.ID
+	if r.ID > 0 {
+		err := db.Delete(&Ref{}, "id=?", r.ID).Error
+		if err != nil && err != gorm.ErrRecordNotFound {
+			return err
+		}
+	}
+	r.IsTemp = true
+	r.ID = 0
+	err := db.Create(r).Error
+
+	if err != nil {
+		return err
+	}
+
+	if toUpdateFileStat {
+		FileUpdated(ctx, prevID, r.ID)
+	}
+
+	return nil
 }
 
 func UpdateIsTemp(ctx context.Context, allocationID, path string, isTemp bool) error {
